latexml_package::engine::base_functions
Function insert_block

pub fn insert_block(
    document: &mut Document,
    contents: &Digested,
    block_attr: HashMap<String, String>,
) -> Result<Vec<Node>>

This attempts to be a generalize vbox construction;

The idea is to receeive block-like material, possibly wrapped in appropriate container which gets attributes.

The contents are constructed in an ltx:CaptureBlock element, designed to accept all reasonable block material from several levels, and then determine which container element is most apprpriate for both the conent & context from block, logical-block or sectional-block, or the inline- variants.