latexml_package::prelude::arena

Struct SymHashMap

pub struct SymHashMap<T>(pub HashMap<SymbolU32, T, FxBuildHasher>);

A convenience abstraction over a String-keyed HashMap

typically used for HashMap<String,Stored> states. The goal is to support both a string interface, as well as the interned tickets interface, while avoiding String allocations internally.

Tuple Fields
0: HashMap<SymbolU32, T, FxBuildHasher>
Implementations

impl<T> SymHashMap<T>

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn get(&self, key: &str) -> Option<&T>

pub fn get_sym(&self, key: SymbolU32) -> Option<&T>

pub fn get_mut(&mut self, key: &str) -> Option<&mut T>

pub fn get_mut_sym(&mut self, key: SymbolU32) -> Option<&mut T>

pub fn contains_key(&self, key: &str) -> bool

pub fn contains_key_sym(&self, key: &SymbolU32) -> bool

pub fn insert(&mut self, key: &str, value: T)

pub fn insert_sym(&mut self, key: SymbolU32, value: T)

pub fn remove(&mut self, key: &str)

pub fn remove_sym(&mut self, key: SymbolU32)

pub fn keys(&self) -> Keys<'_, SymbolU32, T>

pub fn entry(&mut self, key: &str) -> Entry<'_, SymbolU32, T>

pub fn entry_sym(&mut self, key: SymbolU32) -> Entry<'_, SymbolU32, T>

pub fn iter(&self) -> Iter<'_, SymbolU32, T>