latexml_package::prelude::binding

Module content

Structs

FindFileOptions

configuration for searching for a file in the local filesystem

InputDefinitionOptions

a configuration for loading LaTeX definition files (such as .sty, .cls, and their bindings)

InputOptions

configuration for input of a TeX source (content files mostly)

RequireOptions

Functions

build_invocation

Creates a single Tokens representing a TeX invocation of the lead token over a list of arguments.

convert_latex_args

Convert a LaTeX-style argument spec to our Package form. Ie. given $nargs and $optional, being the two optional arguments to something like \newcommand, convert it to the form we use

digest_if

digest_literal

digest_text

find_file

search for a file as prescribed by a FindFileOptions configuration

input

This is essentially the \input equivalent

input_content

Input for cases when the file (or data) is plain TeX material that is expected to contribute content to the document (as opposed to pure definitions).

input_definitions

TODO: Flesh out with the full infrastructure, incremental functionality for now.

install_tag

load_binding

loads a binding from the main binding dispatcher, if available+found

load_class

load_external_binding

loads a binding from an external binding dispatcher, if available+found

load_font_map

load_tex_content

merge_font

preload_font_map

process_options

require_package

An opinionated binding for \RequirePackage.

require_resource

select_relaxng_schema

Selects the RelaxNG schema defining the XML output language