latexml_package::prelude::binding::def::dialect

Function def_conditional

pub fn def_conditional(
    cs: Token,
    paramlist: Option<Parameters>,
    test: Option<Rc<dyn Fn(Vec<ArgWrap>) -> Result<bool, Error>>>,
    options: ConditionalOptions,
) -> Result<(), Error>

Define a conditional control sequence.

Its processing takes place in the Gullet. The test is applied to the arguments (if any), which determines which branch is executed. If the test is undefined, the conditional is a “user defined” one; Two additional primitives are defined \footrue and \foofalse; the test is then determined by the most recently called of those.

If you supply a skipper instead of a test, it is also applied to the arguments and should skip to the right place in the following \or, \else, \fi.