latexml_package::prelude::definition
Module conditional

Conditionals Control sequence definitions. These represent the control sequences for conditionals, as well as \else, \or and \fi.

Structs

Conditional

A Conditional definition; Expandable.

ConditionalOptions

configurations for a conditional.

IfFrame

A Frame of data for the currently active conditional, stored in State

Enums

ConditionalType

classify the standard pieces of a conditional