latexml_package::prelude
Enum DigestedData

pub enum DigestedData {
    TBox(RefCell<Tbox>),
    Whatsit(RefCell<Whatsit>),
    Alignment(RefCell<Alignment>),
    List(RefCell<List>),
    Postponed(Tokens),
    KeyVals(KeyVals),
    RegisterValue(RegisterValue),
    Comment(Comment),
}

These are all kinds of data which we consider officially supported as outputs from the digestion phase of TeX, i.e. from invoking a token.

Variants

TBox(RefCell<Tbox>)

A TeX Box

Whatsit(RefCell<Whatsit>)

A TeX Whatsit (with interior mutability, for setters invoked while stored in state)

Alignment(RefCell<Alignment>)

A TeX Alignment (with interior mutability, for setters invoked while stored in state)

List(RefCell<List>)

A list of Digested data

Postponed(Tokens)

Raw Tokens that were postponed to the digestion phase uninvoked/undigested

KeyVals(KeyVals)

A LaTeX-like digested key-value map

RegisterValue(RegisterValue)

A TeX-like RegisterValue (e.g. a Dimension or Glue)

Comment(Comment)

A TeX comment