latexml_package::prelude
Function build_invocation

pub fn build_invocation<T>(
    token: T,
    args: Vec<Option<Tokens>>,
) -> Result<Tokens, Error>
where
    T: Into<Token>,

Creates a single Tokens representing a TeX invocation of the lead token over a list of arguments.

Note: currently this is near the Mouth representation of data, and deals purely with tokens. A more generic version of this method may be able to support ArgWrap for the argument list. Indeed, the return type may also be lifted to a new generic ArgWrap::Invocation, if there was benefit.