latexml_package::prelude::util
Module pathname

helper methods for file system paths

Structs

PathnameFindOptions

configuration for filesystem search

Functions

absolute

convert a (possibly relative) file path to an absolute one

candidate_pathnames

It’s presumably cheep to concatinate all the pathnames, relative to the cost of testing for files, and this simplifies overall.

canonical

This likely needs portability work!!! (particularly regarding urls, separators, …) AND, care about symbolic links and collapsing ../ !!!

concat

combine a directory and a base name into a full path

cwd

returns the current working directory

directory

obtain the directory portion of a pathname (via Path::parent)

extension

obtain the extension portion of a pathname (via Path::extension)

file_name

transform to a canonical file name, via Path::file_name

file_stem

transform to a base name (via Path::file_stem)

find

find the requested pathname using the options search configuration

is_absolute

absolute paths start with the filesystem root - check if this is one

is_literaldata

checks if the path starts with the “literal:” protocol

is_nasty

check if pathname contains dangerous pieces

is_reloadable

check whether a pathname is reloadable as a TeX definition

is_url

checks if the path is a conforming URL string

kpsewhich

search for a list of candidate names via the external kpsewhich utility returning the first path that is found

protocol

Note that this returns ONLY recognized protocols!

split

Split the pathname into components (dir,name,type). If pathname is absolute, dir starts with volume or ‘/’

url_split

Simple logic for splitting a URL into protocol://base/path