latexml_package::prelude::util
Module text

helpers for extracting structured data from replacement (and other) strings

Enums

Delimiter

kinds of delimiters used by the binding replacement strings

Functions

extract_bracketed

Extract a bracketed subexpression at the start of a larger string. defaults to () delimiters.

trim_end_in_place

trim_in_place

trim_start_in_place