Crate latexml_package

Binding definitions for the LaTeXML converter, reimplemented in Rust

Modules

engine

macros

package

prelude

Generic support layer for any LaTeXML binding file

Macros

AddToCounter

AddToMacro

AssignCatcode

AssignMapping

AssignMeaning

AssignValue

BeginItemize

CounterValue

DeclareFontMap

DeclareOption

DefAccent

DefColumnType

DefConditional

DefConstructor

DefEnvironment

DefEnvironmentIWO

DefEnvironmentWO

DefKeyVal

Defines a new KeyVal Parameter in the given keyset, key and with optional prefix option.prefix. For descriptions of further parameters, see keyval::define.

DefLigature

DefMacro

Defines a macro, a binding analog to \def

DefMath

DefParameterType

DefParameterTypeWO

DefPrimitive

Define a primitive control sequence.

DefRegister

Defines a register with value as the initial value

Digest

DigestIf

DigestText

Dimension

DocType

Expand

Return tokens with all tokens expanded

FindFile

GetKeyVal

GetKeyVals

Glue

InnerPool

Loader shorthand for pool dependencies

InputDefinitions

InstallDefinition

Invocation

Builds a representation of a single command sequence invoked on a Vec<Token> of its arguments. A leading string argument is interpreted as T_CS.

IsDefinable

IsDefined

IsDefinedToken

Let

LoadClass

LoadDefinitions

LoadPool

LookupBool

LookupCatcode

LookupColor

LookupDefinition

LookupFont

LookupMeaning

LookupNumber

LookupRegister

LookupRegisterOrDefault

LookupString

LookupTokens

MergeFont

Merge the current font with the style specifications

MuGlue

NewCounter

ProcessOptions

RawTeX

RefStepCounter

convenience macro for api::counter_dialect::ref_step_counter

RefStepID

convenience macro for api::counter_dialect::ref_step_id

RegisterDocumentNamespace

RegisterNamespace

RelaxNGSchema

RequirePackage

RequireResource

ResetCounter

Revert

StepCounter

Tag

TeX

Today

Tokenize

Tokenize($string); Tokenizes the string using the standard cattable, returning a LaTeXML::Core::Tokens

TokenizeInternal

TokenizeInternal($string); Tokenizes the string using the internal cattable, returning a LaTeXML::Core::Tokens

TypedColumntype

TypedConditional

TypedMacro

TypedPrimitive

XEquals

compile_expansion

Macro for compiling string macro expansions into closures Approach borrowed from diesel-codegen

compile_prototype

Macro for compiling string literal prototypes into a Token and Parameters structs

compile_prototype_for_typed_columntype

Macro for compiling string binding prototypes into ColumnType closures Approach borrowed from diesel-codegen

compile_prototype_for_typed_conditional

Macro for compiling string binding prototypes into Conditional closures

compile_prototype_for_typed_macro

Macro for compiling string binding prototypes into Expandable closures Approach borrowed from diesel-codegen

compile_prototype_for_typed_primitive

Macro for compiling string binding prototypes into Primitive closures

compile_replacement

Macro for compiling string construction replacements into closures that execute the needed ops in libxml. Approach borrowed from diesel-codegen

compile_tokenize

Macro for compiling string literal tokens into their usual Tokens representation

compile_tokenize_internal

Macro for compiling string literal tokens into their preamble Tokens representation

defi_math

defi_opts

defi_register

load_model

parameter_rust_type

parse_prototype

A macro that uses latexml_codegen to expand a prototype at compile time, and initialize all parameters with the in-scope state::at runtime. Returns a (cs, parameters) pair.

Functions

dispatch