latexml_package::prelude

Module token

The Token and its Catcode

Structs

Token

The core immutable syntactic primitive resulting from TeX’s read-in and expansion process We allow the fields to be public, so that we can use builder macros such as

Enums

Catcode

A Token category code, as in TeX

Statics

TOKEN_ALIGN

constant for an ALIGN “&” token

TOKEN_BEGIN

constant for a BEGIN “{” token

TOKEN_CR

constant for a CR “\n” token

TOKEN_END

constant for an END “}” token

TOKEN_ENDCSNAME

constant for T_CS(“\endcsname”)

TOKEN_EXPANDAFTER

constant for T_CS(“\expandafter”)

TOKEN_MATH

constant for a MATH “$” token

TOKEN_PARAM

constant for a PARAM “#” token

TOKEN_RELAX

constant for T_CS(“\relax”)

TOKEN_SPACE

constant for a SPACE “ “ token

TOKEN_SUB

constant for a SUB “_” token

TOKEN_SUPER

constant for a SUPER “^” token